security updates
* no more eval()
* full support for passing/verifying auth tokens
* control what methods are accessible for each collection

+++ collectionapi.js
@@ -1,26 +1,33 @@
 function CollectionAPI(options) {
-
   var self = this;
 
   self._url = __meteor_bootstrap__.require('url');
+  self._querystring = __meteor_bootstrap__.require('querystring');
   self._collections = {};
   self.options = {
     sslEnabled: false,
     listenPort: 3005,
     listenHost: undefined,
+    authToken: undefined,
     privateKeyFile: 'privatekey.pem',
     certificateFile: 'certificate.pem'
   };
   _.extend(self.options, options || {});
 
-  self.addCollection = function() {
-
+  self.addCollection = function(collection, path, options) {
+    collectionOptions = {}
+    collectionOptions[path] = {
+      collection: collection,
+      options: options || {}
+    };
+    _.extend(self._collections, collectionOptions);
   };
 
   self.start = function() {
-    var httpServer, httpOptions;
+    var httpServer, httpOptions, scheme;
 
     if (self.options.sslEnabled === true) {
+      scheme = 'https://';
       httpServer = __meteor_bootstrap__.require('https');
       var fs = __meteor_bootstrap__.require('fs');
 
@@ -29,6 +36,7 @@ function CollectionAPI(options) {
         cert: fs.readFileSync(self.options.certificateFile)
       };
     } else {
+      scheme = 'http://';
       httpServer = __meteor_bootstrap__.require('http');
     }
 
@@ -36,60 +44,82 @@ function CollectionAPI(options) {
     self._server.addListener('request', function(request, response) { self._requestListener(request, response) });
     self._server.listen(self.options.listenPort, self.options.listenHost);
 
-    var protocol = self.options.sslEnabled ? 'https://' : 'http://';
-    var host = self.options.listenHost || 'localhost';
-    console.log('Collection API running on ' +  protocol + host + ':' + self.options.listenPort);
+    console.log('Collection API running on ' +  scheme + (self.options.listenHost || 'localhost') + ':' + self.options.listenPort);
   }
 
   self._requestListener = function(request, response) {
     self._request = request;
     self._response = response;
 
-    // [1] should be the collection path variable name
+    self._requestUrl = self._url.parse(self._request.url);
+
+    // Check for the X-Auth-Token header or auth-token in the query string
+    self._requestAuthToken = self._request.headers['x-auth-token'] ? self._request.headers['x-auth-token'] : self._querystring.parse(self._requestUrl.query)['auth-token'];
+
+    // [1] should be the collection path name
     // [2] is the _id of that collection (optional)
-    self._requestInfo = self._url.parse(self._request.url).pathname.split('/');
+    self._requestPath = self._requestUrl.pathname.split('/');
     
-    // Make sure it looks like a variable name. I know this does not cover
-    // ALL allowed variable names... but since I do an eval, I feel like
-    // I have to check for at least something so it can be considered safe'ish.
-    if (! self._requestInfo[1].match(/^[a-zA-Z_$][0-9a-zA-Z_$]*$/)) {
-      return self._notFoundResponse('Invalid Collection Name');
+    self._requestCollection = self._collections[self._requestPath[1]] ? self._collections[self._requestPath[1]].collection : undefined;
+
+    if (!self._authenticate()) {
+      return self._unauthorizedResponse('Invalid/Missing Auth Token');
     }
 
-    // eval() is icky, but I don't know how else to accomplish this.
-    try {
-      self._requestCollection = eval(self._requestInfo[1]);
-    } catch (e) {
-      self._requestCollection = undefined;
-    }
-
-    // Do our best to ensure we have a Meteor.Collection object
-    if (self._requestCollection === undefined || typeof(self._requestCollection) !== 'object' || self._requestCollection._collection === undefined) {
+    if (!self._requestCollection) {
       return self._notFoundResponse('Collection Object Not Found');
     }
 
     return self._handleRequest();
   };
 
+  self._authenticate = function() {
+    var collectionOptions = self._requestCollectionOptions();
+
+    // Check the collection's auth token
+    if (collectionOptions && collectionOptions.authToken) {
+      return self._requestAuthToken === collectionOptions.authToken;
+    }
+
+    // Check the global auth token
+    if (self.options.authToken) {
+      return self._requestAuthToken === self.options.authToken;
+    }
+  }
+
   self._handleRequest = function() {
+
+    if (!self._requestMethodAllowed(self._request.method)) {
+      return self._notSupportedResponse();
+    }
 
     switch (self._request.method) {
       case 'GET':
-        self._getRequest();
-        break;
+        return self._getRequest();
       case 'POST':
-        self._postRequest();
-        break;
+        return self._postRequest();
       case 'PUT':
-        self._putRequest();
-        break;
+        return self._putRequest();
       case 'DELETE':
-        self._deleteRequest();
-        break;
+        return self._deleteRequest();
       default:
-        self._notSupportedResponse();
+        return self._notSupportedResponse();
     }
 
+  }
+
+  self._requestCollectionOptions = function() {
+    return self._collections[self._requestPath[1]] ? self._collections[self._requestPath[1]].options : undefined;
+  }
+
+  self._requestMethodAllowed = function (method) {
+    var collectionOptions = self._requestCollectionOptions();
+
+    if (collectionOptions && collectionOptions.methods) {
+      return _.indexOf(collectionOptions.methods, method) >= 0;
+    }
+
+    return true;
   }
 
   self._getRequest = function() {
@@ -98,8 +128,8 @@ function CollectionAPI(options) {
 
       try {
         // TODO: A better way to do this?
-        var collection_result = self._requestInfo[2] !== undefined
-            ? self._requestCollection.find(self._requestInfo[2])
+        var collection_result = self._requestPath[2] !== undefined
+            ? self._requestCollection.find(self._requestPath[2])
             : self._requestCollection.find();
 
         var records = [];
@@ -123,7 +153,7 @@ function CollectionAPI(options) {
 
   self._putRequest = function() {
 
-    if (! self._requestInfo[2]) {
+    if (! self._requestPath[2]) {
       return self._notFoundResponse('Missing _id');
     }
 
@@ -136,7 +166,7 @@ function CollectionAPI(options) {
     req.on('end', function() {
       Fiber(function() {
         try {
-          self._requestCollection.update(self._requestInfo[2], JSON.parse(requestData));
+          self._requestCollection.update(self._requestPath[2], JSON.parse(requestData));
         } catch (e) {
           return self._internalServerErrorResponse(e);
         }
@@ -147,13 +177,13 @@ function CollectionAPI(options) {
   }
 
   self._deleteRequest = function() {
-    if (! self._requestInfo[2]) {
+    if (! self._requestPath[2]) {
       return self._notFoundResponse('Missing _id');
     }
 
     Fiber(function() {
       try {
-        self._requestCollection.remove(self._requestInfo[2]);
+        self._requestCollection.remove(self._requestPath[2]);
       } catch (e) {
         return self._internalServerErrorResponse(e);
       }
@@ -171,11 +201,11 @@ function CollectionAPI(options) {
     self._request.on('end', function() {
       Fiber(function() {
         try {
-          self._requestInfo[2] = self._requestCollection.insert(JSON.parse(requestData));
+          self._requestPath[2] = self._requestCollection.insert(JSON.parse(requestData));
         } catch (e) {
           return self._internalServerErrorResponse(e);
         }
-        return self._createdResponse(JSON.stringify({_id: self._requestInfo[2]}));
+        return self._createdResponse(JSON.stringify({_id: self._requestPath[2]}));
       }).run();
     });
   }
@@ -190,6 +220,10 @@ function CollectionAPI(options) {
 
   self._notSupportedResponse = function() {
     self._sendResponse(501, '');
+  }
+
+  self._unauthorizedResponse = function(body) {
+    self._sendResponse(401, JSON.stringify({message: body.toString()}));
   }
 
   self._notFoundResponse = function(body) {
